+++ packages/iov-keycontrol/src/userprofile.ts
@@ -165,16 +165,16 @@ export class UserProfile {
   }
 
   /** Assigns a label to one of the identities in the wallet with the given ID in the primary keyring */
-  public setIdentityLabel(walletId: WalletId, identity: PublicIdentity, label: string | undefined): void {
-    this.primaryKeyring().setIdentityLabel(walletId, identity, label);
+  public setIdentityLabel(identity: PublicIdentity, label: string | undefined): void {
+    this.primaryKeyring().setIdentityLabel(identity, label);
   }
 
   /**
    * Gets the local label of one of the identities in the wallet with the given ID
    * in the primary keyring
    */
-  public getIdentityLabel(id: WalletId, identity: PublicIdentity): string | undefined {
-    const wallet = this.findWalletInPrimaryKeyring(id);
+  public getIdentityLabel(identity: PublicIdentity): string | undefined {
+    const wallet = this.findWalletInPrimaryKeyringByIdentity(identity);
     return wallet.getIdentityLabel(identity);
   }
 
@@ -197,12 +197,11 @@ export class UserProfile {
    * creator field specifies the keypair to be used for signing.
    */
   public async signTransaction(
-    id: WalletId,
     transaction: UnsignedTransaction,
     codec: TxCodec,
     nonce: Nonce,
   ): Promise<SignedTransaction> {
-    const wallet = this.findWalletInPrimaryKeyring(id);
+    const wallet = this.findWalletInPrimaryKeyringByIdentity(transaction.creator);
 
     const { bytes, prehashType } = codec.bytesToSign(transaction, nonce);
     const signature: FullSignature = {
@@ -219,7 +218,6 @@ export class UserProfile {
   }
 
   public async appendSignature(
-    id: WalletId,
     identity: PublicIdentity,
     originalTransaction: SignedTransaction,
     codec: TxCodec,
@@ -229,7 +227,7 @@ export class UserProfile {
       throw new Error("Signing identity's chainId does not match the transaction's chainId");
     }
 
-    const wallet = this.findWalletInPrimaryKeyring(id);
+    const wallet = this.findWalletInPrimaryKeyringByIdentity(identity);
 
     const { bytes, prehashType } = codec.bytesToSign(originalTransaction.transaction, nonce);
     const newSignature: FullSignature = {
@@ -273,6 +271,17 @@ export class UserProfile {
     const wallet = keyring.getWallet(id);
     if (!wallet) {
       throw new Error(`Wallet of id '${id}' does not exist in keyring`);
+    }
+
+    return wallet;
+  }
+
+  private findWalletInPrimaryKeyringByIdentity(identity: PublicIdentity): ReadonlyWallet {
+    const keyring = this.primaryKeyring();
+
+    const wallet = keyring.getWalletByIdentity(identity);
+    if (!wallet) {
+      throw new Error(`No wallet for identity '${JSON.stringify(identity)}' found in keyring`);
     }
 
     return wallet;