+++ packages/iov-lisk/src/liskconnection.spec.ts
@@ -403,12 +403,7 @@ describe("LiskConnection", () => {
             amount: devnetDefaultAmount,
           };
 
-          const signedTransaction = await profile.signTransaction(
-            wallet.id,
-            sendTx,
-            liskCodec,
-            generateNonce(),
-          );
+          const signedTransaction = await profile.signTransaction(sendTx, liskCodec, generateNonce());
 
           const result = await connection.postTx(liskCodec.bytesToPost(signedTransaction));
           await result.blockInfo.waitFor(info => !isBlockInfoPending(info));
@@ -522,7 +517,7 @@ describe("LiskConnection", () => {
         amount: devnetDefaultAmount,
       };
 
-      const signedTransaction = await profile.signTransaction(wallet.id, sendTx, liskCodec, generateNonce());
+      const signedTransaction = await profile.signTransaction(sendTx, liskCodec, generateNonce());
       const bytesToPost = liskCodec.bytesToPost(signedTransaction);
 
       const connection = await LiskConnection.establish(devnetBase);
@@ -550,12 +545,7 @@ describe("LiskConnection", () => {
           amount: devnetDefaultAmount,
         };
 
-        const signedTransaction = await profile.signTransaction(
-          wallet.id,
-          sendTx,
-          liskCodec,
-          generateNonce(),
-        );
+        const signedTransaction = await profile.signTransaction(sendTx, liskCodec, generateNonce());
         const bytesToPost = liskCodec.bytesToPost(signedTransaction);
 
         const connection = await LiskConnection.establish(devnetBase);
@@ -601,7 +591,7 @@ describe("LiskConnection", () => {
         amount: devnetDefaultAmount,
       };
 
-      const signedTransaction = await profile.signTransaction(wallet.id, sendTx, liskCodec, generateNonce());
+      const signedTransaction = await profile.signTransaction(sendTx, liskCodec, generateNonce());
       const bytesToPost = liskCodec.bytesToPost(signedTransaction);
 
       const connection = await LiskConnection.establish(devnetBase);
@@ -634,7 +624,7 @@ describe("LiskConnection", () => {
       };
 
       // Encode creation timestamp into nonce
-      const signedTransaction = await profile.signTransaction(wallet.id, sendTx, liskCodec, generateNonce());
+      const signedTransaction = await profile.signTransaction(sendTx, liskCodec, generateNonce());
 
       const corruptedSignature = signedTransaction.primarySignature.signature.map((x, i) =>
         // tslint:disable-next-line:no-bitwise
@@ -911,9 +901,9 @@ describe("LiskConnection", () => {
         };
 
         const [nonceA, nonceB, nonceC] = await connection.getNonces({ pubkey: sender.pubkey }, 3);
-        const signedA = await profile.signTransaction(wallet.id, sendA, liskCodec, nonceA);
-        const signedB = await profile.signTransaction(wallet.id, sendB, liskCodec, nonceB);
-        const signedC = await profile.signTransaction(wallet.id, sendC, liskCodec, nonceC);
+        const signedA = await profile.signTransaction(sendA, liskCodec, nonceA);
+        const signedB = await profile.signTransaction(sendB, liskCodec, nonceB);
+        const signedC = await profile.signTransaction(sendC, liskCodec, nonceC);
         const bytesToPostA = liskCodec.bytesToPost(signedA);
         const bytesToPostB = liskCodec.bytesToPost(signedB);
         const bytesToPostC = liskCodec.bytesToPost(signedC);
@@ -981,7 +971,7 @@ describe("LiskConnection", () => {
         };
 
         const nonce = await connection.getNonce({ pubkey: sender.pubkey });
-        const signed = await profile.signTransaction(wallet.id, send, liskCodec, nonce);
+        const signed = await profile.signTransaction(send, liskCodec, nonce);
         const bytesToPost = liskCodec.bytesToPost(signed);
 
         const postResult = await connection.postTx(bytesToPost);
@@ -1037,7 +1027,7 @@ describe("LiskConnection", () => {
         };
 
         const nonce = await connection.getNonce({ pubkey: sender.pubkey });
-        const signed = await profile.signTransaction(wallet.id, send, liskCodec, nonce);
+        const signed = await profile.signTransaction(send, liskCodec, nonce);
         const bytesToPost = liskCodec.bytesToPost(signed);
 
         const postResult = await connection.postTx(bytesToPost);