+++ packages/iov-rise/src/riseconnection.spec.ts
@@ -391,12 +391,7 @@ describe("RiseConnection", () => {
             amount: defaultSendAmount,
           };
 
-          const signedTransaction = await profile.signTransaction(
-            wallet.id,
-            sendTx,
-            riseCodec,
-            generateNonce(),
-          );
+          const signedTransaction = await profile.signTransaction(sendTx, riseCodec, generateNonce());
 
           const result = await connection.postTx(riseCodec.bytesToPost(signedTransaction));
           await result.blockInfo.waitFor(info => !isBlockInfoPending(info));
@@ -485,7 +480,7 @@ describe("RiseConnection", () => {
         amount: defaultSendAmount,
       };
 
-      const signedTransaction = await profile.signTransaction(wallet.id, sendTx, riseCodec, generateNonce());
+      const signedTransaction = await profile.signTransaction(sendTx, riseCodec, generateNonce());
       const bytesToPost = riseCodec.bytesToPost(signedTransaction);
 
       const connection = await RiseConnection.establish(base);
@@ -507,12 +502,7 @@ describe("RiseConnection", () => {
           amount: defaultSendAmount,
         };
 
-        const signedTransaction = await profile.signTransaction(
-          wallet.id,
-          sendTx,
-          riseCodec,
-          generateNonce(),
-        );
+        const signedTransaction = await profile.signTransaction(sendTx, riseCodec, generateNonce());
         const bytesToPost = riseCodec.bytesToPost(signedTransaction);
 
         const connection = await RiseConnection.establish(base);
@@ -557,7 +547,7 @@ describe("RiseConnection", () => {
         amount: defaultSendAmount,
       };
 
-      const signedTransaction = await profile.signTransaction(wallet.id, sendTx, riseCodec, generateNonce());
+      const signedTransaction = await profile.signTransaction(sendTx, riseCodec, generateNonce());
 
       const corruptedSignature = signedTransaction.primarySignature.signature.map((x, i) =>
         // tslint:disable-next-line:no-bitwise
@@ -703,8 +693,8 @@ describe("RiseConnection", () => {
         };
 
         const [nonceA, nonceB] = await connection.getNonces({ pubkey: sender.pubkey }, 2);
-        const signedA = await profile.signTransaction(wallet.id, sendA, riseCodec, nonceA);
-        const signedB = await profile.signTransaction(wallet.id, sendB, riseCodec, nonceB);
+        const signedA = await profile.signTransaction(sendA, riseCodec, nonceA);
+        const signedB = await profile.signTransaction(sendB, riseCodec, nonceB);
         const bytesToPostA = riseCodec.bytesToPost(signedA);
         const bytesToPostB = riseCodec.bytesToPost(signedB);
 
@@ -765,7 +755,7 @@ describe("RiseConnection", () => {
         };
 
         const nonce = await connection.getNonce({ pubkey: sender.pubkey });
-        const signed = await profile.signTransaction(wallet.id, send, riseCodec, nonce);
+        const signed = await profile.signTransaction(send, riseCodec, nonce);
         const bytesToPost = riseCodec.bytesToPost(signed);
 
         const postResult = await connection.postTx(bytesToPost);
@@ -819,7 +809,7 @@ describe("RiseConnection", () => {
         };
 
         const nonce = await connection.getNonce({ pubkey: sender.pubkey });
-        const signed = await profile.signTransaction(wallet.id, send, riseCodec, nonce);
+        const signed = await profile.signTransaction(send, riseCodec, nonce);
         const bytesToPost = riseCodec.bytesToPost(signed);
 
         const postResult = await connection.postTx(bytesToPost);